NOT TESTABLE YET Update 05/12: Re-implemented the ListView of the participants in another way. Re-implemented the SMS reading, but still having problems here. Re-implemented contacts comparison.
05/10: Started coding three User Stories instead of one.
US014: Because I need a fragment to work on, I created the fragment to show participants of an event. (MockUp 10/25 Rückmeldungen_zugesagt)
I implemented a swipeToRefresh layout to trigger an update on the participants. The code should be okay here. (US043)

While coding the SMSparsing, I directly implemented a contact list comparison which was a matter of 10 lines of code. (US015)